Better handling of existing entry data
Previously Categories Field didn’t play nice with entries with existing
category data. Now it does. If you add the categories field to an
existing project and ask it to sync with native categories then the
field will load native category data into the field display and also
when replace front end tags. This ensures that the field behaves ‘as
expected’ when implemented within existing projects with lots of
existing entries and category data.